Revert "tdf#96505 Get rid of cargo cult "long" integer literals"
This partly reverts commit 8ae33b1652cb1e654c426350169d3bb9fa031a4f.

Added a comment why 1L is needed in this file.

Reviewed-on: https://gerrit.libreoffice.org/30259
Tested-by: jan iversen <jani@documentfoundation.org>

+++ cppcanvas/source/mtfrenderer/implrenderer.cxx
@@ -1577,8 +1577,10 @@ namespace cppcanvas
                         // #i44110# correct null-sized output - there
                         // are metafiles which have zero size in at
                         // least one dimension
-                        const Size aMtfSizePix( ::std::max( (int)aMtfSizePixPre.Width(), 1 ),
-                                                ::std::max( (int)aMtfSizePixPre.Height(), 1 ) );
+                      
+                        // Remark the 1L cannot be replaced, that would cause max to compare long/int
+                        const Size aMtfSizePix( ::std::max( aMtfSizePixPre.Width(), 1L ),
+                                                ::std::max( aMtfSizePixPre.Height(), 1L ) );
 
                         // Setup local transform, such that the
                         // metafile renders itself into the given
@@ -2914,8 +2916,9 @@ namespace cppcanvas
 
             // #i44110# correct null-sized output - there are shapes
             // which have zero size in at least one dimension
-            const Size aMtfSizePix( ::std::max( (int)aMtfSizePixPre.Width(), 1 ),
-                                    ::std::max( (int)aMtfSizePixPre.Height(), 1 ) );
+            // Remark the 1L cannot be replaced, that would cause max to compare long/int
+            const Size aMtfSizePix( ::std::max( aMtfSizePixPre.Width(), 1L ),
+                                    ::std::max( aMtfSizePixPre.Height(), 1L ) );
 
             sal_Int32 nCurrActions(0);
             ActionFactoryParameters aParms(aStateStack,
